Add new installation paths for setup dependencies
go-bindata was coming from a lot of unmaintained sources so I decided to upgrade it
to the one maintained source with a recent commit and a way to install it via brew.
That however also meant that the go-bindata path changed from the GOPATH to /usr/local/bin,
so I had to make a couple of changes to reflect that.